+++ internal/kosync/api_syncs.go
@@ -9,18 +9,21 @@ package kosync
 import (
 	"fmt"
 
+	"git.obth.eu/atjontv/kosync/internal/legacy"
 	"github.com/gofiber/fiber/v2"
 )
 
 func (app *Kosync) SyncsPostProgress(c *fiber.Ctx) error {
 	// Parse payload
-	var data DocumentData
+	var data legacy.DocumentData
 	if err := c.BodyParser(&data); err != nil {
 		return err
 	}
+	doc := DocumentDataToNew(&data, c.Locals("current_user_id").(string))
 
-	app.PrintDebug("Syncs", c.Locals("requestid").(string), fmt.Sprintf("User '%s' sent progress for document '%s'", c.Locals("current_user").(string), data.Document))
-	if err := app.AddOrUpdateDocument(c.Locals("current_user").(string), data); err != nil {
+	app.PrintDebug("Syncs", c.Locals("requestid").(string), fmt.Sprintf("User '%s' sent progress for document '%s'", c.Locals("current_user_name").(string), data.Document))
+	//if err := app.LegacyDb.AddOrUpdateDocument(c.Locals("current_user_name").(string), data); err != nil {
+	if err := app.Db.CreateOrUpdateDocument(&doc); err != nil {
 		return err
 	}
 
@@ -32,13 +35,17 @@ func (app *Kosync) SyncsGetProgress(c *fiber.Ctx) error {
 	if documentId == "-" {
 		return fiber.ErrNotFound
 	}
-	app.PrintDebug("Syncs", c.Locals("requestid").(string), fmt.Sprintf("User '%s' requested progress of document '%s'", c.Locals("current_user").(string), documentId))
+	app.PrintDebug("Syncs", c.Locals("requestid").(string), fmt.Sprintf("User '%s' requested progress of document '%s'", c.Locals("current_user_name").(string), documentId))
 
 	// Find document
-	docData, found := app.Db.Users[c.Locals("current_user").(string)].Documents[documentId]
+	//docData, found := app.LegacyDb.GetUser(c.Locals("current_user_name").(string)).Documents[documentId]
+	docData, found, err := app.Db.FindDocumentById(c.Locals("current_user_id").(string), documentId)
+	if err != nil {
+		return err
+	}
 	if !found {
 		return fiber.ErrNotFound
 	}
 
-	return c.JSON(DocumentData{ProgressData: docData.ProgressData, Document: documentId})
+	return c.JSON(FileDataFromNew(docData))
 }