+++ internal/kosync/api_webui.go
@@ -10,29 +10,42 @@ import (
 	"encoding/base64"
 	"encoding/json"
 
+	"git.obth.eu/atjontv/kosync/internal/legacy"
 	"github.com/gofiber/fiber/v2"
 )
 
 type UiDocumentData struct {
 	Id string `json:"id"`
-	FileData
-	History []FileData `json:"history"`
+	legacy.FileData
+	History []legacy.FileData `json:"history"`
 }
 
 func (app *Kosync) ApiGetDocumentsAll(c *fiber.Ctx) error {
-	data, found := app.Db.Users[c.Locals("current_user").(string)]
+	user, found, err := app.Db.FindUserByUsername(c.Locals("current_user_name").(string))
+	if err != nil {
+		return err
+	}
 	if !found {
 		return fiber.ErrNotFound
 	}
 
-	result := make([]UiDocumentData, 0, len(data.Documents))
-	for id, doc := range data.Documents {
-		history, found := data.History[doc.DocumentId]
-		if !found {
-			result = append(result, UiDocumentData{id, doc, make([]FileData, 0)})
-		} else {
-			result = append(result, UiDocumentData{id, doc, history.DocumentHistory})
+	docs, err := app.Db.AllDocumentsOfUser(user.Id)
+	if err != nil {
+		return err
+	}
+	result := make([]UiDocumentData, 0, len(docs))
+	for i := range docs {
+		history, err := app.Db.GetDocumentHistory(user.Id, docs[i].Id)
+		if err != nil {
+			continue
 		}
+
+		docHistory := make([]legacy.FileData, 0, len(history))
+		for j := range history {
+			docHistory = append(docHistory, FileDataFromNew(&history[j]))
+		}
+
+		result = append(result, UiDocumentData{Id: docs[i].Id, FileData: FileDataFromNew(&docs[i]), History: docHistory})
 	}
 
 	c.Set("Access-Control-Allow-Origin", "*")
@@ -40,12 +53,18 @@ func (app *Kosync) ApiGetDocumentsAll(c *fiber.Ctx) error {
 }
 
 func (app *Kosync) ApiPutDocument(c *fiber.Ctx) error {
+	user, _, err := app.Db.FindUserByUsername(c.Locals("current_user_name").(string))
+	if err != nil {
+		return err
+	}
+
 	var document UiDocumentData
 	if err := c.BodyParser(&document); err != nil {
 		return err
 	}
 
-	if err := app.UpdateDocumentPrettyName(c.Locals("current_user").(string), document.DocumentId, document.PrettyName); err != nil {
+	doc := FileDataToNew(&document.FileData, user.Id)
+	if err := app.Db.CreateOrUpdateDocument(&doc); err != nil {
 		return err
 	}
 
@@ -53,7 +72,10 @@ func (app *Kosync) ApiPutDocument(c *fiber.Ctx) error {
 }
 
 func (app *Kosync) ApiAuthBasic(c *fiber.Ctx) error {
-	user := app.Db.Users[c.Locals("current_user").(string)]
+	user, _, err := app.Db.FindUserByUsername(c.Locals("current_user_name").(string))
+	if err != nil {
+		return err
+	}
 	type UserData struct {
 		Username string `json:"username"`
 		Key      string `json:"key"`