+++ internal/legacy/database_models.go
@@ -1,10 +1,10 @@
 //
-// File:        internal/kosync/database_models.go
+// File:        internal/legacy/database_models.go
 // Project:     https://git.obth.eu/atjontv/kosync
 // Copyright:   © 2025-2026 Thomas Obernosterer. Licensed under the EUPL-1.2 or later
 //
 
-package kosync
+package legacy
 
 type Database struct {
 	Schema int                 `json:"schema"`