+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
@@ -22,12 +21,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [2026.04.1] - 2026-04-27
+## [2026.05.0] - 2026-02-05
+
+### Added
+- SQLite Database (named `kosync.db`) with automatic migration from JSON to SQLite
+- `.env` Configuration File (named `kosync.env`)
+
+### Changed
+
+### Deprecated
+- `database.json` Database file
+- Backup creation and restore (can no longer be done via CLI)
+
+## [2026.04.1] - 2026-01-27
 
 ## Fix
 - Pretty Name got lost during progress push
 
-## [2026.04.0] - 2026-04-27
+## [2026.04.0] - 2026-01-27
 
 ## Added
 - WebUI to view documents and history